+++ client/src/diagnostics/ValidatorExecutable.ts
@@ -20,12 +20,12 @@ export class ValidatorExecutable extends Validator {
     constructor(path: string, public syntax: string, public customPattern: string) { super(path); }
 
     validate(domainInfo: DomainInfo, problemFiles: ProblemInfo[], onSuccess: (diagnostics: Map<string, Diagnostic[]>) => void, onError: (error: string) => void): void {
-        let domainFilePath = Util.toPddlFile("domain", domainInfo.getText());
+        let domainFilePath = Util.toPddlFileSync("domain", domainInfo.getText());
 
         let diagnostics = this.createEmptyDiagnostics(domainInfo, problemFiles);
 
         if (!problemFiles.length) {
-            let problemFilePath = Util.toPddlFile("problem", PddlFactory.createEmptyProblem('dummy', domainInfo.name));
+            let problemFilePath = Util.toPddlFileSync("problem", PddlFactory.createEmptyProblem('dummy', domainInfo.name));
             let pathToUriMap: [string, string][] = [[domainFilePath, domainInfo.fileUri]];
 
             this.validateOneProblem(domainFilePath, problemFilePath, output => {
@@ -35,7 +35,7 @@ export class ValidatorExecutable extends Validator {
         }
         else {
             problemFiles.forEach(problemFile => {
-                let problemFilePath = Util.toPddlFile("problem", problemFile.getText());
+                let problemFilePath = Util.toPddlFileSync("problem", problemFile.getText());
                 let pathToUriMap: [string, string][] = [[domainFilePath, domainInfo.fileUri], [problemFilePath, problemFile.fileUri]];
 
                 // todo: the issues in the domain file should only be output once, not as many times as there are problem files
@@ -67,7 +67,7 @@ export class ValidatorExecutable extends Validator {
 
                 let pathUriTuple = pathToUriMap.find(tuple => tuple[0] == pattern.getFilePath(match))
 
-                if (!pathUriTuple) continue; // this is not a file of interest
+                if (!pathUriTuple) { continue; } // this is not a file of interest
 
                 let uri = pathUriTuple[1];
                 let diagnostic = new Diagnostic(pattern.getRange(match), pattern.getMessage(match),Validator.toSeverity(pattern.getSeverity(match)));