+++ client/src/planning/PlannerExecutable.ts
@@ -27,10 +27,10 @@ export class PlannerExecutable extends Planner {
         super(plannerPath, plannerOptions);
     }
 
-    plan(domainFileInfo: DomainInfo, problemFileInfo: ProblemInfo, planParser: PddlPlanParser, parent: PlannerResponseHandler): Promise<Plan[]> {
+    async plan(domainFileInfo: DomainInfo, problemFileInfo: ProblemInfo, planParser: PddlPlanParser, parent: PlannerResponseHandler): Promise<Plan[]> {
 
-        let domainFilePath = Util.toPddlFile("domain", domainFileInfo.getText());
-        let problemFilePath = Util.toPddlFile("problem", problemFileInfo.getText());
+        let domainFilePath = await Util.toPddlFile("domain", domainFileInfo.getText());
+        let problemFilePath = await Util.toPddlFile("problem", problemFileInfo.getText());
 
         let command = this.plannerSyntax.replace('$(planner)', Util.q(this.plannerPath))
             .replace('$(options)', this.plannerOptions)
@@ -65,7 +65,7 @@ export class PlannerExecutable extends Planner {
                     }
 
                     let plans = planParser.getPlans();
-                    resolve(plans);
+                    resolve(plans); // todo: should we resolve() even if we reject()ed above?
                     thisPlanner.child = null;
                 });
 
@@ -77,8 +77,8 @@ export class PlannerExecutable extends Planner {
             thisPlanner.child.stderr.on('data', (data: any) => parent.handleOutput("Error: " + data));
 
             thisPlanner.child.on("close", (code: any, signal: any) => {
-                if (code) console.log("Exit code: " + code);
-                if (signal) console.log("Exit Signal: " + signal);
+                if (code) { console.log("Exit code: " + code); }
+                if (signal) { console.log("Exit Signal: " + signal); }
             });
         });
     }